lenochka-memory/requirements.txt
# Lenochka Memory — dependencies
sentence-transformers>=3.0
sqlite-vec>=0.1
numpy>=1.24
requests>=2.28
python-dotenv>=1.0
